moved context files to a context directory under src
